Adds a ProcessSelector and updates validations and logging
This change add the Sensor::ProcessSelector class which is
responsible for starting the appropriate processs for
writing, updating, deleting, or execution based on inputs.
When a filename is passed with no flags the default behavior
will be execution if the file passes validations.

The Sensor::CommandLineValidator has been updated to ensure
a file exists if the options specify anything other than
writing. The framework expects that full paths will be provided
for updating, deletion, and execution for the time being.

Logging has been updated to include process and user info
when using the #activity method.