Introduces HttpForwarder and updates logging
The Sensor::HttpForwarder establishes a TCPSocket connection with
localhost:2000 unless ADDRESS and PORT variables are set in
the environment. The forwarder sends a JSON payload as an HTTP
POST over the TCP connection. Sensor ships with a TCPServer for
testing in spec/tcp_test_server.rb.

This change also makes some simplifications to application logging,
moving most the responsibility of creating JSON for the logs into
the Sensor::HttpForwarder and Sensor::ProcessSelector.