Add LogFormatting, exception handling, and testing
Adds the helper module Sensor::LogFormatting to help with
formatting process start times and usernames returned from the
system.

Introduces exception handling in Sensor::ProcessSelector to
give the user feedback when permissions are incorrect and in
Sensor::HttpForwarder to hand connetion refused errors.

Creates a docker-compose.yml file to allow running the spec suite
in a Linux environment by runing `docker-compose up` from the
home directory of the Gem.